Add Header logic and setup Event driver in future
This commit does a few things. First it checks the headers for a few
specific things:

1. That it's UserAgent is GitHub-Hookshot
2. That it has a GitHub webhook event associated with it

If both of those conditions are true the code then parses the event into
the Event type for use later. If any of the things prior to this fail it
returns a 400 request. If successful this then Event will eventually be
used to get Thearesia to do things on the repo.

For now these events just print out the JSON but in the future they will
contain the functions for Thearesia to execute on the repos it's
associated with.