Abstract new issue code to sync module
We had most of the code originally in main in order to deal with
prototyping. Now though we want to encapsulate that into a module
and define functions for each of the different kinds of syncing that
can occur. This commit moves that code to a new module sync and cleans
up the main.rs file to incorporate that, as well as aliasing the Result
type to make things look a bit cleaner in our code as well.